use hdrhistogram to collect duration data for cache operations
We want a (potentially) richer amount of information for cache
operations than just "the average", and the way to do that is capturing
a histogram of duration data rather than just computing the average when
stats are requested.  For better or worse, `hdrhistogram`
is (apparently) the most mature histogram package found on crates.io, so
let's go with it.