fix parsing for -dependency_file
This option was getting passed as a preprocessor argument and the
additional arguments that it requires (e.g. `-MT`) were getting passed
as dependency arguments.  So when we went to call the preprocessor, the
compiler complained about the missing arguments.  Instead, we should be
classifying it as specifying a dependency target, which will direct it
to the correct set of arguments.