fix parsing for -dependency-file
This option was getting passed as a preprocessor argument and the
additional arguments that it requires (e.g. `-MT`) were getting passed
as dependency arguments.  So when we went to call the preprocessor, the
compiler complained about the missing arguments.  Instead, we should be
classifying it as specifying a dependency target, which will direct it
to the correct set of arguments.

We also take this opportunity to move it to be a clang-exclusive option;
the same option exists in GCC, but it's Darwin-specific and unlikely to
be actually used there.